dom/base/test/jsmodules/importmaps/test_importMap_with_external_script.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test speculative preload of external script doesn't conflict with import map</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<!--
    These tests check that speculative preloading, which could happen before
    the import map is installed, doesn't load the wrong modules.
-->

<script type="importmap">
  {
    "imports": {
      "bare": "./good/module_0.js",
      "./bad/module_1.js": "./good/module_1.js",
      "./bad/module_2.js": "./good/module_2.js",
      "./bad/module_3.js": "./good/module_3.js"
    }
  }
</script>

<!--
Test bareword import (not supported before import map installed).
-->
<script type="module" src="module_importMap_with_external_script_0.js"></script>

<!--
Test mapping from missing resource to existing resource (not found before
import map installed).
-->
<script type="module" src="module_importMap_with_external_script_1.js"></script>

<!--
Test mapping from one existing resource to another (would load wrong resource before
import map installed).
-->
<script type="module" src="module_importMap_with_external_script_2.js"></script>

<!--
Test mapping from one existing resource to another with circular dependency.
-->
<script type="module" src="module_importMap_with_external_script_3.js"></script>

<script>
  SimpleTest.waitForExplicitFinish();

  let passCount = 0;
  const expectedCount = 4;

  function success(name) {
    ok(true, "Test passed, loaded " + name);
    passCount++;
    if (passCount == expectedCount) {
      SimpleTest.finish();
    }
  }
</script>
<body></body>
